+++ Tests/GRDBTests/DatabasePoolReleaseMemoryTests.swift
@@ -205,7 +205,7 @@ class DatabasePoolReleaseMemoryTests: GRDBTestCase {
             }
         }
         
-        var cursor: ColumnCursor<Int>? = nil
+        var cursor: FastDatabaseValueCursor<Int>? = nil
         do {
             try! makeDatabasePool().write { db in
                 try db.execute("CREATE TABLE items (id INTEGER PRIMARY KEY)")